Create Tree#height
This method takes one argument, a node, and uses recursion to find the number of edges in the longest path from the given node to a leaf node.

It recursively finds the height of the left path and the right path of the node and selects the longer of the two.